backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.1
psycopg2-binary>=2.9.9
django-cors-headers>=4.3.1
openai>=1.30.0
celery>=5.4.0
redis>=5.0.4
requests>=2.32.0
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
Pillow>=10.3.0

opencv-python-headless>=4.10.0.84
flake8>=7.1.1
